projects/helloworld/api/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.43.1
anyio==4.8.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
frozenlist==1.5.0
google-auth==2.37.0
gotrue==2.11.1
griffe==1.5.5
groq==0.15.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpath-python==1.0.6
logfire-api==3.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.3.1
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.59.8
packaging==24.2
postgrest==0.19.1
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-ai==0.0.19
pydantic-ai-slim==0.0.19
pydantic-graph==0.0.19
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.1.0
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==13.1
yarl==1.18.3
